Port UI to Avalonia
